# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScorpStudy by Bishal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes for the current user
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, title]
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /quiz-results:
    get:
      operationId: listQuizResults
      tags: [history]
      summary: List quiz results for the current user
      responses:
        "200":
          description: List of quiz results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizResult"
    post:
      operationId: saveQuizResult
      tags: [history]
      summary: Save a quiz result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizResultInput"
      responses:
        "201":
          description: Saved quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /flashcard-sets:
    get:
      operationId: listFlashcardSets
      tags: [history]
      summary: List saved flashcard sets
      responses:
        "200":
          description: List of flashcard sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardSet"
    post:
      operationId: saveFlashcardSet
      tags: [history]
      summary: Save a flashcard set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardSetInput"
      responses:
        "201":
          description: Saved flashcard set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
  /flashcard-sets/{id}:
    delete:
      operationId: deleteFlashcardSet
      tags: [history]
      summary: Delete a flashcard set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /summaries:
    get:
      operationId: listSummaries
      tags: [history]
      summary: List saved summaries
      responses:
        "200":
          description: List of summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Summary"
    post:
      operationId: saveSummary
      tags: [history]
      summary: Save a summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryInput"
      responses:
        "201":
          description: Saved summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
  /images:
    get:
      operationId: listImages
      tags: [history]
      summary: List generated images
      responses:
        "200":
          description: List of images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedImage"
    post:
      operationId: saveImage
      tags: [history]
      summary: Save a generated image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratedImageInput"
      responses:
        "201":
          description: Saved image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
  /images/{id}:
    delete:
      operationId: deleteImage
      tags: [history]
      summary: Delete an image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /mindmaps:
    get:
      operationId: listMindmaps
      tags: [history]
      summary: List saved mind maps
      responses:
        "200":
          description: List of mind maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mindmap"
    post:
      operationId: saveMindmap
      tags: [history]
      summary: Save a mind map
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MindmapSaveInput"
      responses:
        "201":
          description: Saved mind map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mindmap"
  /translations:
    get:
      operationId: listTranslations
      tags: [history]
      summary: List translation history
      responses:
        "200":
          description: List of translations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Translation"
    post:
      operationId: saveTranslation
      tags: [history]
      summary: Save a translation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslationInput"
      responses:
        "201":
          description: Saved translation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Translation"
  /translations/{id}:
    delete:
      operationId: deleteTranslation
      tags: [history]
      summary: Delete a translation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Flashcard:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
      required:
        - front
        - back
...
    MindmapNode:
      type: object
      properties:
        label:
          type: string
        children:
          type: array
          items:
            type: string
      required:
        - label
        - children
    MindmapResponse:
      type: object
      properties:
        center:
          type: string
        branches:
          type: array
          items:
            $ref: "#/components/schemas/MindmapNode"
      required:
        - center
        - branches
...
    Note:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - content
        - tags
        - createdAt
        - updatedAt
    NoteInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        userId:
          type: string
      required:
        - title
        - userId
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
...
    QuizResult:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        topic:
          type: string
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: number
        difficulty:
          type: string
        questionType:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - score
        - total
        - percentage
        - createdAt
    QuizResultInput:
      type: object
      properties:
        topic:
          type: string
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: number
        difficulty:
          type: string
        questionType:
          type: string
        userId:
          type: string
      required:
        - score
        - total
        - percentage
        - userId
    FlashcardSet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        topic:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - topic
        - cards
        - createdAt
    FlashcardSetInput:
      type: object
      properties:
        topic:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        userId:
          type: string
      required:
        - topic
        - cards
        - userId
    Summary:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        sourceType:
          type: string
        originalText:
          type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        examQuestions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - summary
        - keyPoints
        - examQuestions
        - createdAt
    SummaryInput:
      type: object
      properties:
        sourceType:
          type: string
        originalText:
          type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        examQuestions:
          type: array
          items:
            type: string
        userId:
          type: string
      required:
        - summary
        - userId
    GeneratedImage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        prompt:
          type: string
        enhancedPrompt:
          type: ["string", "null"]
        imageUrl:
          type: string
        style:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - prompt
        - imageUrl
        - createdAt
    GeneratedImageInput:
      type: object
      properties:
        prompt:
          type: string
        enhancedPrompt:
          type: ["string", "null"]
        imageUrl:
          type: string
        style:
          type: string
        userId:
          type: string
      required:
        - prompt
        - imageUrl
        - userId
    Mindmap:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        topic:
          type: string
        mapData:
          $ref: "#/components/schemas/MindmapResponse"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - topic
        - mapData
        - createdAt
    MindmapSaveInput:
      type: object
      properties:
        topic:
          type: string
        mapData:
          $ref: "#/components/schemas/MindmapResponse"
        userId:
          type: string
      required:
        - topic
        - mapData
        - userId
    Translation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        originalText:
          type: string
        translatedText:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - originalText
        - translatedText
        - targetLanguage
        - createdAt
    TranslationInput:
      type: object
      properties:
        originalText:
          type: string
        translatedText:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string
        userId:
          type: string
      required:
        - originalText
        - translatedText
        - targetLanguage
        - userId
...