# lib/api-zod/src/generated/types/flashcardSet.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScorpStudy by Bishal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from './flashcard';

export interface FlashcardSet {
  id: number;
  userId: string;
  topic: string;
  cards: Flashcard[];
  createdAt: Date;
}