# lib/api-zod/src/generated/types/mindmap.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScorpStudy by Bishal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MindmapResponse } from './mindmapResponse';

export interface Mindmap {
  id: number;
  userId: string;
  topic: string;
  mapData: MindmapResponse;
  createdAt: Date;
}